requirements.txt
flask==3.0.0
python-dotenv>=1.0.0
waitress
sqlalchemy
flask-sqlalchemy
flask-login
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
shapely==2.0.1
colorama==0.4.6
reportlab==4.0.4
pdfkit==1.0.0
numpy==1.24.3
werkzeug==3.0.1
colorama==0.4.6
reportlab==4.0.4
pdfkit==1.0.0
numpy==1.24.3
werkzeug==3.0.1